feat: added pods list in admin side changes using tsx file which supports html tags for table view formatting
